% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{addCaptionToPlotly}
\alias{addCaptionToPlotly}
\title{Add a caption to a plotly object}
\usage{
addCaptionToPlotly(pl, caption)
}
\arguments{
\item{pl}{a \code{\link[plotly]{plotly}} object}

\item{caption}{String with caption.}
}
\value{
The updated \code{\link[plotly]{plotly}} object
with a caption included
}
\description{
Add a caption to a plotly object
}
\author{
Laure Cougnaud
}
