% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-utility-caption-subtitle.R
\name{addLabsToPlotly}
\alias{addLabsToPlotly}
\title{Add labels, i.e. caption of subtitle to a plotly object}
\usage{
addLabsToPlotly(
  pl,
  caption = NULL,
  subtitle = NULL,
  legend = FALSE,
  legendPosition = "right"
)
}
\arguments{
\item{pl}{a \code{\link[plotly]{plotly}} object}

\item{caption}{String with caption.}

\item{subtitle}{String with subtitle}

\item{legend}{Logical, if TRUE (by default)
a legend is available in the plot.}

\item{legendPosition}{String with position of the legend,
'right' by default.}
}
\value{
The updated \code{\link[plotly]{plotly}} object
with a caption included
}
\description{
Add labels, i.e. caption of subtitle to a plotly object
}
\author{
Laure Cougnaud
}
