% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-staticScatterplotClinData.R
\name{addLayerToScatterPlot}
\alias{addLayerToScatterPlot}
\title{Helper function to add layer to scatter plot}
\usage{
addLayerToScatterPlot(
  gg,
  aesVar,
  pars,
  generalPars,
  layerFunction,
  useHandlers = FALSE
)
}
\arguments{
\item{gg}{\code{\link[ggplot2]{ggplot}} object}

\item{aesVar}{layers specific aesthetics list of layer specific aesthetics}

\item{pars}{list of parameters specific to the layer \code{\link[ggplot2]{aes}}}

\item{generalPars}{overall, not layer specific parameters can be overwritten by \code{pars}}

\item{layerFunction}{function to use for adding the layer e.g. \code{\link[ggplot2]{geom_line}}}

\item{useHandlers}{if \code{TRUE} we use handlers to repress the expected warning:
\code{Hover is set via the 'text' aesthetic in ggplot}, we need to pass this aesthetic to have it available
in plotly even though it is not used by geom_point. Defaults to \code{FALSE}}
}
\value{
\code{\link[ggplot2]{ggplot}} object
}
\description{
Helper function to add layer to scatter plot
}
\author{
Adriaan Blommaert Laure Cougnaud
}
