% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-render-html.R
\name{buildBook}
\alias{buildBook}
\title{Build the book}
\usage{
buildBook(htmlFiles)
}
\arguments{
\item{htmlFiles}{character vector with path to HTML files}
}
\value{
String with path to the front page of the 
report.
}
\description{
Build the book
}
\author{
Laure Cougnaud
}
