% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-render_medicalMonitoringReport.R
\name{checkReportTitles}
\alias{checkReportTitles}
\title{Check report titles}
\usage{
checkReportTitles(configFiles, configDir = "./config")
}
\arguments{
\item{configFiles}{Character vector with config file names}

\item{configDir}{String with directory with config files,
('config' by default)}
}
\value{
A named vector with the report titles and the corresponding config file
}
\description{
Check uniqueness of report titles across the config files.
If not unique titles are provided, an error is returned.
}
\author{
Michela Pasetto
}
