% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-render_clinDataReviewReport.R
\name{checkTemplatesName}
\alias{checkTemplatesName}
\title{Checks of config files template.}
\usage{
checkTemplatesName(
  configFiles,
  configDir = file.path(inputDir, "config"),
  inputDir = "."
)
}
\arguments{
\item{configFiles}{Character vector with name or path of the config file(s).}

\item{configDir}{String with directory with config files,
by default a 'config' folder in \code{inputDir}.\cr
It should contain a general 'config.yml' file and dedicated
'config-[X].yml' for each chapter.
The order of each chapter is specified in the 'config' slot in the general 
general 'config.yml'.}

\item{inputDir}{String with input directory,
working directory by default.}
}
\value{
Updated \code{configFiles}
}
\description{
Check if the templates specified in the input config files
don't originate from multiple sources (e.g. custom and R package
via the parameter \code{templatePackage}).
If so, the corresponding config files are not considered.
}
\author{
Laure Cougnaud
}
