% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-render_clinDataReviewReport.R
\name{clinDataReview-common-args-report}
\alias{clinDataReview-common-args-report}
\title{Common parameters for the clinical data reporting function}
\arguments{
\item{indexPath}{String with path to the index file,
by default 'index.Rmd' in \code{inputDir}.}

\item{configDir}{String with directory with config files,
by default a 'config' folder in \code{inputDir}.\cr
It should contain a general 'config.yml' file and dedicated
'config-[X].yml' for each chapter.
The order of each chapter is specified in the 'config' slot in the general 
general 'config.yml'.}

\item{inputDir}{String with input directory,
working directory by default.}

\item{outputDir}{String with output directory,
('report' by default).}

\item{intermediateDir}{String with intermediate directory ('interim'
by default), where
markdown files and rds file specifying Js libraries (with \code{knit_meta}) for
each sub report are stored.}
}
\value{
No return value, used for the documentation of 
the clinical data reporting functions of the package.
}
\description{
Common parameters for the clinical data reporting function
}
