% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-utility.R
\name{combineButtonsAndPlot}
\alias{combineButtonsAndPlot}
\title{Combine select box(es) and the plot}
\usage{
combineButtonsAndPlot(x)
}
\arguments{
\item{x}{Object of class \code{clinDataReview}}
}
\value{
\code{x} object:
\itemize{
\item{with the \code{plot} element containing a combination of the 
\code{buttons} and the \code{plot}}
\item{without the \code{buttons} element}
}
}
\description{
Combine select box(es) and the plot
}
