% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-convertHtmlToAspx.R
\name{convertReportToAspx}
\alias{convertReportToAspx}
\title{Convert report format from .html to .aspx}
\usage{
convertReportToAspx(reportDir = ".")
}
\arguments{
\item{reportDir}{String for the path to the directory where
the clinical data reports are stored, defaults to current directory}
}
\value{
no return value, files in directory are modified
}
\description{
Report files generated as output of \code{\link{render_clinDataReviewReport}}  are converted from .html to 
to .aspx format by changing extensions and cross-links of all files within the directory \code{reportDir}.
This allows for deployment on SharePoint.
}
