% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{countNLines}
\alias{countNLines}
\title{Count number of lines in a vector}
\usage{
countNLines(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Integer vector of length \code{x} with number
of lines
}
\description{
Count number of lines in a vector
}
\examples{
countNLines(x = c("A\nB", "blabla", "This\nis\na\nsentence."))
}
\author{
Laure Cougnaud
}
