% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-skeleton.R
\name{createClinDataReviewReportSkeleton}
\alias{createClinDataReviewReportSkeleton}
\title{Create the skeleton of a report}
\usage{
createClinDataReviewReportSkeleton(dir = ".")
}
\arguments{
\item{dir}{String with the path of the directory where 
the skeleton should be created. The current working
directory is used as default.}
}
\value{
The files to run a report are written in the specified 
directory. To run the report, the user can call the 
\code{\link{render_clinDataReviewReport}}.
}
\description{
Creates the skeleton of a report to start running the analyses.
}
\details{
This function is meant to get familiar with the use of the package and the 
necessary files to create a report.

It will create a ready-to-use report with example data from the \code{clinUtils} 
package. After getting use to the file structure, the user can 
substitute the example data with 
custom data sets and add specific configuration files.
}
