% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-skeleton.R
\name{createExampleMetadata}
\alias{createExampleMetadata}
\title{Create an example metadata file}
\usage{
createExampleMetadata(dir)
}
\arguments{
\item{dir}{String, path to the directory.}
}
\value{
Nothing, the example metadata file is created in the specified 
directory.
}
\description{
Create an example of metadata file for the \code{\link{createClinDataReviewReportSkeleton}}.
}
