% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-skeleton.R
\name{createMainConfigSkeleton}
\alias{createMainConfigSkeleton}
\title{Create the config file for the skeleton}
\usage{
createMainConfigSkeleton(dirName, dirData)
}
\arguments{
\item{dirName}{String, path to the directory.}

\item{dirData}{String, path to the directory of the data.}
}
\description{
This function creates the main config file for the \code{\link{reportSkeleton}} 
with the directory where the data are stored.
}
