% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-templates.R
\name{createTemplateDoc}
\alias{createTemplateDoc}
\title{Create documentation for clinical data template reports
available in the 'template' folder of the package.}
\usage{
createTemplateDoc(
  templatePath = system.file("template", package = "clinDataReview")
)
}
\arguments{
\item{templatePath}{string with path where the template Rmd reports
and associated JSON schema files are stored,
by default path of the installed version of the package.
This parameter is only for expert use of the package.}
}
\value{
Character vector with Rd code containing description
for all template documents.
}
\description{
If a JSON schema file available, the information relative
to the template is extracted from this file with the function
\code{JSONSchToRd}.
}
\references{
\href{https://json-schema.org/understanding-json-schema/}{JSON schema specification}
}
\author{
Laure Cougnaud
}
