% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-render_clinDataReviewReport.R
\name{exportSessionInfoToMd}
\alias{exportSessionInfoToMd}
\title{Combine all session informations across all clinical data reports
and export them into a dedicated Markdown document}
\usage{
exportSessionInfoToMd(sessionInfos, mdFiles, intermediateDir = "interim")
}
\arguments{
\item{sessionInfos}{List with \code{\link{sessionInfo}} objects}

\item{mdFiles}{Character vector with Markdown files}

\item{intermediateDir}{String with intermediate directory ('interim'
by default), where
markdown files and rds file specifying Js libraries (with \code{knit_meta}) for
each sub report are stored.}
}
\value{
String with path to Markdown file containing the session information,
NULL if no session information(s) are provided.
}
\description{
Combine all session informations across all clinical data reports
and export them into a dedicated Markdown document
}
\author{
Laure Cougnaud
}
