% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{formatHoverText}
\alias{formatHoverText}
\title{Format hover text for use in plotly interactive plots.
The labels are wrapped to multiple lines if exceed the width of the plotly
hover box, e.g. in case labels for points with same x/y coordinates overlap,
and corresponding labels are truncated.}
\usage{
formatHoverText(x, label, width = 50)
}
\arguments{
\item{x}{Vector with hover text information.}

\item{label}{Label for the variable}

\item{width}{Integer, number of characters at 
which the hover text should be cut at to multiple lines.}
}
\value{
String with formatted hover label.
}
\description{
Format hover text for use in plotly interactive plots.
The labels are wrapped to multiple lines if exceed the width of the plotly
hover box, e.g. in case labels for points with same x/y coordinates overlap,
and corresponding labels are truncated.
}
\author{
Laure Cougnaud
}
