% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-utility-axisLabels.R
\name{getAxisLab}
\alias{getAxisLab}
\title{Get axis label}
\usage{
getAxisLab(axisVar, axisLab, labVar = NULL, data, labelVars)
}
\arguments{
\item{axisVar}{String with column of \code{data} containing variable for the axis.}

\item{axisLab}{(optional) String with label for \code{axisVar}}

\item{labVar}{String with column of \code{data} containing variable(s) whose
content should be displayed in the axis (with its labels).}

\item{data}{Data.frame with data.}

\item{labelVars}{Named character vector containing variable labels.}
}
\value{
String with label for the axis.
}
\description{
Get axis label
}
