% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-utility-axisLabels.R
\name{getAxisLabs}
\alias{getAxisLabs}
\title{Set different variables for the x-axis labels}
\usage{
getAxisLabs(data, var, labVars)
}
\arguments{
\item{data}{Data.frame with data.}

\item{var}{String with variable displayed
in the axis.}

\item{labVars}{Character vector with variable(s) to be displayed 
as the labels of the ticks in the axis.}
}
\value{
Named character vector.
The names are the position in the x-axis,
the values are the new labels.
}
\description{
Set different variables for the x-axis labels
}
\author{
Laure Cougnaud
}
