% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-utility-facets.R
\name{getDimGgplot}
\alias{getDimGgplot}
\title{Get plot dimensions}
\usage{
getDimGgplot(gg)
}
\arguments{
\item{gg}{\code{\link[ggplot2]{ggplot}}}
}
\value{
Numeric vector with number of rows ('nrow')
and columns ('ncol') of the plot
}
\description{
Get plot dimensions
}
\author{
Laure Cougnaud
}
