% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-render-utility.R
\name{getExtraDirs}
\alias{getExtraDirs}
\title{Get extra directory(ies) required for the clinical data review report}
\usage{
getExtraDirs(inputDir = ".", configDir = file.path(inputDir, "config"))
}
\arguments{
\item{inputDir}{String with input directory,
working directory by default.}

\item{configDir}{String with directory with config files,
by default a 'config' folder in \code{inputDir}.\cr
It should contain a general 'config.yml' file and dedicated
'config-[X].yml' for each chapter.
The order of each chapter is specified in the 'config' slot in the general 
general 'config.yml'.}
}
\value{
Character vector with extra directories required by the report
}
\description{
By default, the 'figures', 'tables' and patient profiles folders 
(\code{patientProfilePath} parameter in the general
config file, if specified) in the input directory are considered.
}
\author{
Laure Cougnaud
}
