% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation-utility.R
\name{getFctCode}
\alias{getFctCode}
\title{Get function code}
\usage{
getFctCode(fct)
}
\arguments{
\item{fct}{a R function}
}
\value{
String with function code
}
\description{
Get function code
}
