% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-utility-dimensions.R
\name{getHeightCaption}
\alias{getHeightCaption}
\title{Get height of caption}
\usage{
getHeightCaption(caption)
}
\arguments{
\item{caption}{String with caption.}
}
\value{
Integer with height in pixels 
for this element.
}
\description{
Get height of caption
}
\author{
Laure Cougnaud
}
