% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-utility-dimensions.R
\name{getHeightLab}
\alias{getHeightLab}
\title{Get height of labels: title, subtitle or caption}
\usage{
getHeightLab(lab)
}
\arguments{
\item{lab}{String with label.}
}
\value{
Integer with height in pixels 
for this element.
}
\description{
Get height of labels: title, subtitle or caption
}
\author{
Laure Cougnaud
}
