% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-utility-dimensions.R
\name{getHeightLabs}
\alias{getHeightLabs}
\title{Get height of labels of the plot,
i.e caption or subtitle}
\usage{
getHeightLabs(lab)
}
\arguments{
\item{lab}{String with label.}
}
\value{
Integer with height in pixels 
for this element.
}
\description{
Get height of labels of the plot,
i.e caption or subtitle
}
\author{
Laure Cougnaud
}
