% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-utility-dimensions.R
\name{getHeightSubtitle}
\alias{getHeightSubtitle}
\title{Get height of subtitle}
\usage{
getHeightSubtitle(subtitle)
}
\arguments{
\item{subtitle}{String with subtitle}
}
\value{
Integer with height in pixels 
for this element.
}
\description{
Get height of subtitle
}
\author{
Laure Cougnaud
}
