% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-render-html.R
\name{getHTMLToc}
\alias{getHTMLToc}
\title{Get HTML toc}
\usage{
getHTMLToc(toc)
}
\arguments{
\item{toc}{data.frame with TOC info}
}
\value{
Character vector with HTML toc
}
\description{
Get HTML toc
}
