% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-render-html.R
\name{getIndexHTMLTitle}
\alias{getIndexHTMLTitle}
\title{Get index of the line containing the HTML title in a vector of HTML strings}
\usage{
getIndexHTMLTitle(x)
}
\arguments{
\item{x}{Character vector with HTML}
}
\value{
Integer vector with index of the title in the vector \code{x}
}
\description{
Get index of the line containing the HTML title in a vector of HTML strings
}
