% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-errorbarClinData.R
\name{getJitterVar}
\alias{getJitterVar}
\title{Add jitter to the variable of the plot,
based on the different groups of a grouping variable}
\usage{
getJitterVar(data, var, byVar)
}
\arguments{
\item{data}{Data.frame with data.}

\item{var}{String with variable to add a jitter to.}

\item{byVar}{String with variable containing the
groups to jitter by.}
}
\value{
Numeric vector of length \code{nrow(data)}
containing the jittered variable.
}
\description{
Add jitter to the variable of the plot,
based on the different groups of a grouping variable
}
\author{
Laure Cougnaud
}
