% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-utility-dimensions.R
\name{getMargins}
\alias{getMargins}
\title{Get margins for a clinical data plot}
\usage{
getMargins(
  title = NULL,
  subtitle = NULL,
  xLab = NULL,
  caption = NULL,
  facet = FALSE,
  includeLegend = TRUE,
  legendPosition = "right"
)
}
\arguments{
\item{title}{String with title for the plot.}

\item{subtitle}{String with subtitle}

\item{xLab}{String with label for \code{xVar}.}

\item{caption}{String with caption.}

\item{legendPosition}{String with position of the legend,
'right' by default.}
}
\value{
List with margins
}
\description{
Get margins for a clinical data plot
}
\author{
Laure Cougnaud
}
