% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-utility.R
\name{getMdHeader}
\alias{getMdHeader}
\title{Get Markdown header, for creation in Rmarkdown.}
\usage{
getMdHeader(title, level = 1)
}
\arguments{
\item{title}{String with header title.}

\item{level}{Integer of length 1 with header depth/level,
1 by default}
}
\value{
String with Markdown header,
to be included in R within \code{cat}.
}
\description{
The depth is extracted:
\enumerate{
\item from the \code{settings} if a
specified depth is provided in the 'rmd_file_depth' for the
current knitted file
\item \code{level} parameter otherwise
}
}
\seealso{
Other clinical data reporting: 
\code{\link{checkReportTitles}()},
\code{\link{forceParams}()},
\code{\link{getParamsFromConfig}()},
\code{\link{gitbook_clinDataReview_report}()},
\code{\link{html_clinDataReview_report}()},
\code{\link{knitPrintClinDataReview}()},
\code{\link{postProcessReport}()},
\code{\link{render_clinDataReviewReport}()}
}
\concept{clinical data reporting}
