% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-utility-dimensions.R
\name{getPositionAndMargins}
\alias{getPositionAndMargins}
\title{Get margins for a clinical data plot}
\usage{
getPositionAndMargins(
  title = NULL,
  subtitle = NULL,
  xLab = NULL,
  caption = NULL,
  facet = FALSE,
  includeLegend = TRUE,
  legendPosition = "right"
)
}
\arguments{
\item{title}{String with title for the plot.}

\item{subtitle}{String with subtitle}

\item{xLab}{String with label for \code{xVar}.}

\item{caption}{String with caption.}

\item{facet}{Logical, if TRUE the plot
contains facets.}

\item{includeLegend}{Logical, if TRUE (by default)
a legend is available in the plot.}

\item{legendPosition}{String with position of the legend,
'right' by default.}
}
\value{
List with:
\itemize{
\item{'margin': }{List with bottom ('t') and top ('t')
margins in pixels}
\item{'position': }{List with position
of the following plot elements:
\itemize{
\item{on top of the plot: subtitle and legend
(if positioned at the top).\cr
The position is defined as the distance in pixels
from the top of the plotting area to the bottom
of the element (\code{yanchor = 'bottom'})}
\item{at the bottom of the plot: caption, xLab
and legend (if positioned at the bottom).\cr
The position is defined as the distance in pixels
from the bottom of the plotting area to the top
of the element (\code{yanchor = 'top'})\cr
Especially, the legend should be positioned with
anchor 'top' such as the margins are automatically
expanded if the legend contains multiple rows.}
}}
}
}
\description{
Get margins for a clinical data plot
}
\author{
Laure Cougnaud
}
