% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-utility-dimensions.R
\name{getSizePlotClinData}
\alias{getSizePlotClinData}
\title{Get dimensions for a clinical data plot}
\usage{
getSizePlotClinData(
  width = NULL,
  height = NULL,
  gg = NULL,
  nrow = 1L,
  ncol = 1L,
  legend = TRUE,
  legendPosition = "right",
  caption = NULL,
  subtitle = NULL
)
}
\arguments{
\item{width}{Numeric, width of the plot in pixels,
700 by default.}

\item{height}{Numeric, height of the plot in pixels,
700 by default.}

\item{gg}{\code{\link[ggplot2]{ggplot}}}

\item{nrow}{single-length integer specifying the 
number of facet rows in the plot. (default = 1) 
Overwritten if \code{gg} is specified.}

\item{ncol}{single-length integer specifying the 
number of facet columns in the plot. (default = 1)
Overwritten if \code{gg} is specified.}

\item{legend}{Logical, if TRUE (by default)
a legend is available in the plot.}

\item{legendPosition}{String with position of the legend,
'right' by default.}

\item{caption}{String with caption.}

\item{subtitle}{String with subtitle}
}
\value{
Numeric vector with width ('width')
and height ('height') of the plot
in pixels.
}
\description{
This function set sensitive defaults dimensions
for a plot in the package.
This includes:
\itemize{
\item{setting a default width for a figure to 
fit in a standard clinical data review report}
\item{increasing the figure height:}{
\itemize{
\item{for facetted plot, ensuring that each 
facet is relatively squared}
\item{if a caption or a subtitle is specified}
\item{if a legend is set at the bottom or the top
of the plot}
}
\item{increasing the figure width if a legend
is set at the left or the right of the plot}
}
}
}
\author{
Laure Cougnaud
}
