% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-render-html.R
\name{getTocNumbering}
\alias{getTocNumbering}
\title{Get TOC numbering}
\usage{
getTocNumbering(levels)
}
\arguments{
\item{levels}{vector with levels of the section,
in the order as available in the book.}
}
\value{
Character vector with section numbers
}
\description{
Get TOC numbering
}
\author{
Laure Cougnaud
}
