% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-formats.R
\name{html_clinDataReview_report}
\alias{html_clinDataReview_report}
\title{Clinical data format for rmarkdown report.}
\usage{
html_clinDataReview_report(extra_dependencies = NULL, ...)
}
\arguments{
\item{extra_dependencies}{NULL by default.}

\item{...}{Extra parameters passed to the
\link[rmarkdown]{html_document} function.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown]{render}}.
}
\description{
This function only kept for back-compatibility, 
\code{\link[rmarkdown]{html_document}}
can be used instead.
}
\seealso{
Other clinical data reporting: 
\code{\link{checkReportTitles}()},
\code{\link{forceParams}()},
\code{\link{getMdHeader}()},
\code{\link{getParamsFromConfig}()},
\code{\link{gitbook_clinDataReview_report}()},
\code{\link{knitPrintClinDataReview}()},
\code{\link{postProcessReport}()},
\code{\link{render_clinDataReviewReport}()}
}
\author{
Laure Cougnaud
}
\concept{clinical data reporting}
