% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-utility.R
\name{knit_print.clinDataReview}
\alias{knit_print.clinDataReview}
\title{Print \code{clinDataReviewTable} object in a knitted document
(e.g. Rmarkdown document).}
\usage{
\method{knit_print}{clinDataReview}(x, ...)
}
\arguments{
\item{x}{Object of class \code{clinDataReview}}

\item{...}{Extra parameters for compatibility with \code{\link[knitr]{knit_print}},
not used currently.}
}
\description{
Print \code{clinDataReviewTable} object in a knitted document
(e.g. Rmarkdown document).
}
\author{
Laure Cougnaud
}
