% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-utility-layout.R
\name{layoutClinData}
\alias{layoutClinData}
\title{Set layout for a clinical data plot.}
\usage{
layoutClinData(
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  caption = NULL,
  subtitle = NULL,
  includeLegend = FALSE,
  legendPosition = "right",
  facet = FALSE,
  nrow = 1L,
  ncol = 1L,
  width,
  height,
  ...
)
}
\arguments{
\item{xLab}{String with label for \code{xVar}.}

\item{title}{String with title for the plot.}

\item{caption}{String with caption.}

\item{subtitle}{String with subtitle}

\item{legendPosition}{String with position of the legend,
'right' by default.}

\item{facet}{Logical (FALSE by default), 
does the plot contains facets?}

\item{nrow}{single-length integer specifying the 
number of facet rows in the plot. (default = 1) 
Overwritten if \code{gg} is specified.}

\item{ncol}{single-length integer specifying the 
number of facet columns in the plot. (default = 1)
Overwritten if \code{gg} is specified.}

\item{width}{Numeric, width of the plot in pixels,
700 by default.}

\item{height}{Numeric, height of the plot in pixels,
700 by default.}

\item{...}{Any parameters for the \link{code[plotly]{layout}}
function. This should contain at least the plot object.}

\item{pl}{a \code{\link[plotly]{plotly}} object}
}
\value{
The updated \code{\link[plotly]{plotly}} object
}
\description{
Set layout for a clinical data plot.
}
\author{
Laure Cougnaud
}
