% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-render_medicalMonitoringReport.R
\name{medicalMonitoring-common-args-report}
\alias{medicalMonitoring-common-args-report}
\title{Common parameters for the medical monitoring reporting function}
\arguments{
\item{indexPath}{String with path to the index file,
('index.Rmd' by default).}

\item{configDir}{String with directory with config files,
('config' by default).
It should contain a general 'config.yml' file and dedicated
'config-[X].yml' for each chapter.
The order of each chapter is specified in the 'config' slot in the general 
general 'config.yml'.}

\item{outputDir}{String with output directory,
('report' by default).}

\item{intermediateDir}{String with intermediate directory ('interim'
by default), where
markdown files and rds file specifying Js libraries (with \code{knit_meta}) for
each sub report are stored.}
}
\description{
Common parameters for the medical monitoring reporting function
}
