% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-render_clinDataReviewReport.R
\name{merge.sessionInfo}
\alias{merge.sessionInfo}
\title{Merge multiple session information}
\usage{
\method{merge}{sessionInfo}(...)
}
\arguments{
\item{...}{objects of type \code{\link{sessionInfo}}}
}
\value{
\code{\link{sessionInfo}} with combined information
}
\description{
Merge multiple session information
}
\author{
Laure Cougnaud
}
