% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-skeleton.R
\name{moveSkeletonFiles}
\alias{moveSkeletonFiles}
\title{Move skeleton files from the package to a directory}
\usage{
moveSkeletonFiles(dir)
}
\arguments{
\item{dir}{String, path to the directory.}
}
\value{
Nothing, the files are available in the specified 
directory.
}
\description{
This function moves the files used to create the skeleton from the 
package to a specified directory.
}
