% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-skeleton.R
\name{moveXpt}
\alias{moveXpt}
\title{Move data sets from clinUtils}
\usage{
moveXpt(dir)
}
\arguments{
\item{dir}{String, path to the directory.}
}
\value{
Nothing, the data are saved in the dedicated location.
}
\description{
Move SDTM data sets available in \code{clinUtils} into a 
specified local directory.
}
