% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-utility.R
\name{print.clinDataReview}
\alias{print.clinDataReview}
\title{Print a \code{clinDataReview} object in the console}
\usage{
\method{print}{clinDataReview}(x, ...)
}
\arguments{
\item{x}{Object of class \code{clinDataReview}}

\item{...}{Extra parameters for compatibility with \code{\link{print}},
not used currently.}
}
\value{
No returned value, the object is printed into the console.
}
\description{
Print a \code{clinDataReview} object in the console
}
