% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{varToFm}
\alias{varToFm}
\title{Get formula for a specific variable,
to be used in aesthetic specification in \code{\link[plotly]{plot_ly}}.}
\usage{
varToFm(var)
}
\arguments{
\item{var}{Character vector with variable to combine.
Otherwise with the '+' operator.}
}
\value{
\code{\link[stats]{as.formula}}
}
\description{
Get formula for a specific variable,
to be used in aesthetic specification in \code{\link[plotly]{plot_ly}}.
}
\author{
Laure Cougnaud
}
