% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-render-html.R
\name{buildBook}
\alias{buildBook}
\title{Build the book}
\usage{
buildBook(htmlFiles, verbose = TRUE)
}
\arguments{
\item{htmlFiles}{character vector with path to HTML files}

\item{verbose}{Logical, if TRUE (FALSE by default) progress messages are printed
during the report execution.}
}
\value{
String with path to the front page of the 
report.
}
\description{
Build the book
}
\author{
Laure Cougnaud
}
