% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-render_clinDataReviewReport.R
\name{checkReportTitles}
\alias{checkReportTitles}
\title{Check report titles}
\usage{
checkReportTitles(
  configFiles,
  configDir = file.path(inputDir, "config"),
  inputDir = "."
)
}
\arguments{
\item{configFiles}{Character vector with config file names}

\item{configDir}{String with directory with config files,
('config' by default)}

\item{inputDir}{String with input directory,
working directory by default.}
}
\value{
A named vector with the report titles and the corresponding config file
}
\description{
Check uniqueness of report titles across the config files.
If not unique titles are provided, an error is returned.
}
\seealso{
Other clinical data reporting: 
\code{\link{convertMdToHtml}()},
\code{\link{forceParams}()},
\code{\link{getMdHeader}()},
\code{\link{getParamsFromConfig}()},
\code{\link{gitbook_clinDataReview_report}()},
\code{\link{html_clinDataReview_report}()},
\code{\link{knitPrintClinDataReview}()},
\code{\link{render_clinDataReviewReport}()}
}
\author{
Michela Pasetto
}
\concept{clinical data reporting}
