% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-skeleton.R
\name{createMainConfigSkeleton}
\alias{createMainConfigSkeleton}
\title{Create the config file for the skeleton}
\usage{
createMainConfigSkeleton(dir, dirData)
}
\arguments{
\item{dir}{String, path to the directory.}

\item{dirData}{String, path to the directory of the data.}
}
\value{
No return value, a file _config.yml_
is created in the specified directory.
}
\description{
This function creates the main config file for the \code{\link{createClinDataReviewReportSkeleton}} 
with the directory where the data are stored.
}
