% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation-filterData.R
\name{filterData}
\alias{filterData}
\title{Filter dataset based on specified filters.}
\usage{
filterData(
  data,
  filters,
  keepNA = TRUE,
  returnAll = FALSE,
  verbose = FALSE,
  labelVars = NULL,
  labelData = "data"
)
}
\arguments{
\item{data}{Data.frame with data.}

\item{filters}{Unique filter or list of filters.
Each filter should be a list containing:
\itemize{
\item{'var': }{String with variable from \code{data} to filter on.}
\item{'value': }{(optional) Character vector with values from \code{var} to consider.}
\item{'valueFct': }{(optional) Function (or string wit this function)
to be applied on \code{var} to extract value to consider}
\item{'op': }{(optional) String with operator used to retain records from \code{value}.
If not specified, the inclusion operator: '\%in\%' is considered, a.k.a
records with \code{var} in \code{value} are retained.}
\item{'rev': }{(optional) Logical, if TRUE (FALSE by default), 
filtering condition based on \code{value}/\code{valueFct} is reversed.}
\item{'keepNA': }{(optional) Logical, if TRUE (by default), missing values in \code{var}
are retained. If not specified, \code{keepNA} general parameter is used.}
\item{'varsBy': }{(optional) Character vector with variables in \code{data} containing groups to filter by}
\item{'varNew': }{(optional) String with name for the new variable created}
\item{'labelNew': }{(optional) String with label for \code{varNew}}
}
If a list of filters is specified, the logical operator (see \code{\link[base]{Logic}})
linking the different conditions
can be specified between the two conditions, e.g.: 
\code{list(list(var = "SEX", value = "F"), "&", list(var = "COUNTRY", value = "DEU"))}.}

\item{keepNA}{Logical, if TRUE (by default) missing values in \code{var} are retained.
If set to FALSE, missing values are ignored for all filters.
The specification within \code{filters} prevails on this parameter.}

\item{returnAll}{Logical:
\itemize{
\item{if FALSE (by default): }{the \code{data} for only the filtered records
is returned.}
\item{if TRUE: }{the full \code{data} is returned. 
Records are flagged based on the \code{filters} condition, in a new column:
\code{varNew} (if specified), or 'keep' otherwise; containing TRUE
 if the record fulfill all conditions, FALSE otherwise}
}}

\item{verbose}{Logical, if TRUE (FALSE by default) progress messages are printed
in the current console.
For the visualizations, progress messages during download
of subject-specific report are displayed in the browser console.}

\item{labelVars}{Named character vector containing variable labels.}

\item{labelData}{(optional) String with label for input \code{data},
that will be included in progress messages.}
}
\value{
Filtered \code{data} if \code{returnAll} is FALSE (by default).
Otherwise \code{data} with additional column: \code{keep} or \code{varNew} (if specified),
containing TRUE for records which fullfill the specified
condition(s) and FALSE otherwise.
}
\description{
A dataset can be filtered:
\itemize{
\item{based:
\itemize{
\item{on a specific \code{value} of interest)}
\item{on a function of the variable (\code{valueFct} parameter), 
e.g. maximum of the variable)}
\item{to retain only non missing values of a variable (\code{keepNA} set to \code{FALSE})}
}
}
\item{by groups (\code{varsBy} parameter)}
}
**Note that by default, missing values in the filtering variable are retained
(which differs from the default behaviour in R)**.
To filter missing records, please set the \code{keepNA} parameter to \code{FALSE}.
}
\examples{
library(clinUtils)

data(dataADaMCDISCP01)
labelVars <- attr(dataADaMCDISCP01, "labelVars")

dataDM <- dataADaMCDISCP01$ADSL

## single filter

# filter with inclusion criteria:
filterData(
	data = dataDM, 
	filters = list(var = "SEX", value = "M"), 
	verbose = TRUE
)

# filter with non-inclusion criteria
filterData(
	data = dataDM, 
	filters = list(var = "SEX", value = "M", rev = TRUE), 
	verbose = TRUE
)

# filter based on inequality operator
filterData(
	data = dataDM, 
	filters = list(var = "AGE", value = 75, op = "<="), 
	verbose = TRUE
)

# missing values are retained by default!
dataDMNA <- dataDM
dataDMNA[1 : 2, "AGE"] <- NA
filterData(
	data = dataDMNA, 
	filters = list(var = "AGE", value = 75, op = "<="), 
	verbose = TRUE
)

# filter missing values on variable
filterData(
	data = dataDMNA, 
	filters = list(var = "AGE", value = 75, op = "<=", keepNA = FALSE), 
	verbose = TRUE
)

# retain only missing values
filterData(
	data = dataDMNA, 
	filters = list(var = "AGE", value = NA, keepNA = TRUE), 
	verbose = TRUE
)

# filter missing values
filterData(
	data = dataDMNA, 
	filters = list(var = "AGE", keepNA = FALSE), 
	verbose = TRUE
)


## multiple filters

# by default the records fulfilling all conditions are retained ('AND')
filterData(
	data = dataDM, 
	filters = list(
		list(var = "AGE", value = 75, op = "<="),
		list(var = "SEX", value = "M")
	), 
	verbose = TRUE
)

# custom operator:
filterData(
	data = dataDM, 
	filters = list(
		list(var = "AGE", value = 75, op = "<="),
		"|",
		list(var = "SEX", value = "M")
	), 
	verbose = TRUE
)
}
\author{
Laure Cougnaud
}
