% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-templates.R
\name{getPathTemplate}
\alias{getPathTemplate}
\title{Get path of template clinical data report}
\usage{
getPathTemplate(file, package = "clinDataReview")
}
\arguments{
\item{file}{String with name of the template Rmd document}

\item{package}{String, which package the template should be extracted from,
by default the \code{clinDataReview} package.}
}
\value{
String with path to the template in the installed \code{clinDataReview}
package
}
\description{
Get path of template clinical data report
}
\examples{
\donttest{
pathDivisionTemplate <- getPathTemplate("divisionTemplate.Rmd") # get path template in the package
tmp <- file.copy(from = pathDivisionTemplate, to = tempdir()) # copy to temporary dir (for example)
params <- list(reportTitle = "Adverse events")
rmarkdown::render(file.path(tempdir(), basename(pathDivisionTemplate))) # run file 
}
}
\author{
Laure Cougnaud
}
