% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-utility-dimensions.R
\name{getPositionAndMargins}
\alias{getPositionAndMargins}
\title{Get margins for a clinical data plot}
\usage{
getPositionAndMargins(
  title = NULL,
  subtitle = NULL,
  xLab = NULL,
  caption = NULL,
  facet = FALSE,
  includeLegend = TRUE,
  legendPosition = "right"
)
}
\arguments{
\item{title}{String with title for the plot.}

\item{subtitle}{String with subtitle}

\item{xLab}{String with label for \code{xVar}.}

\item{caption}{String with caption.}

\item{legendPosition}{String with position of the legend,
'right' by default.}
}
\value{
List with:
\itemize{
\item{'margin'}{List with bottom ('t') and top ('t')
margins in pixels}
\item{'position': }{List with position
of the following plot elements:
title, subtitle, caption, xLab
of the following elements.\cr
The position is defined in distance in pixelx from the bottom
or the top of the plotting region, depending on the 
location of the element.
}
}
}
\description{
Get margins for a clinical data plot
}
\author{
Laure Cougnaud
}
