% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-utility-facets.R
\name{getSizePlotClinData}
\alias{getSizePlotClinData}
\title{Get dimensions for a clinical data plot}
\usage{
getSizePlotClinData(
  width = NULL,
  height = NULL,
  gg = NULL,
  nrow = 1L,
  ncol = 1L,
  legend = TRUE,
  legendPosition = "right"
)
}
\arguments{
\item{width}{Numeric, width of the plot in pixels,
700 by default.}

\item{height}{Numeric, height of the plot in pixels,
700 by default.}

\item{gg}{\code{\link[ggplot2]{ggplot}}}

\item{nrow}{single-length integer specifying the 
number of facet rows in the plot. (default = 1) 
Overwritten if \code{gg} is specified.}

\item{ncol}{single-length integer specifying the 
number of facet columns in the plot. (default = 1)
Overwritten if \code{gg} is specified.}

\item{legend}{Logical, if TRUE (by default)
a legend is available in the plot.}

\item{legendPosition}{String with position of the legend,
'right' by default.}
}
\value{
Numeric vector with width ('width')
and height ('height') of the plot
in pixels.
}
\description{
Get dimensions for a clinical data plot
}
\author{
Laure Cougnaud
}
